Hide implementation details
